Move Genesis tests to `ouroboros-consensus-diffusion`
